+++ src/colmap/controllers/base_controller.h
